Fix merchant despawning on reload
During world initialization, `world#getBlock()` will return bedrock for
a few ticks until everything is up and running. This was causing the
Merchant to think that his market had disappeared, leading him to
despawn every re-load.

This "solves" that by considering bedrock a valid market. A better
solution would probably be to check if the world is initializing before
permitting bedrock, but I'm unable to find any way to do that.