Fix Merchant AI (#2)
* Unify block references under ModBlocks

`ModBlocks` wasn't being used causing a comparison in
`EntityMerchant#isMarketValid` to always fail.

* Enable Merchant AI

Enable and slightly tweak the merchant's AI. Fixes the pathing target
from being a close, relative distance to being the actual position of
the market.

Implements `isAtMarket` with a 1 block stopping distance.

Fixes a related crash where breaking a market without an associated
merchant would throw an NPE, crashing the game.

* Fix AI targeting

Merchant AI previously targeted the center of the market block, which
kinda worked, but was kind of weird. Now the Merchant will target the
block one block behind the market podium.

Additionally, the merchant will spawn in this same spot instead of
spawning on top of the market.

* Fix merchant despawning on reload

During world initialization, `world#getBlock()` will return bedrock for
a few ticks until everything is up and running. This was causing the
Merchant to think that his market had disappeared, leading him to
despawn every re-load.

This "solves" that by considering bedrock a valid market. A better
solution would probably be to check if the world is initializing before
permitting bedrock, but I'm unable to find any way to do that.